cleanup facet-common
https://coveord.atlassian.net/browse/KIT-3031
